+++ src/analyzing/Dataflow_svalue.ml
@@ -107,7 +107,12 @@ let eq c1 c2 =
   match (c1, c2) with
   | G.Lit l1, G.Lit l2 -> eq_literal l1 l2
   | G.Cst t1, G.Cst t2 -> eq_ctype t1 t2
-  | G.Sym e1, G.Sym e2 -> AST_utils.with_structural_equal G.equal_expr e1 e2
+  | G.Sym e1, G.Sym e2 ->
+      (* We only consider two `Sym`s equal when they are exactly the same. If two
+       * `Sym`s are structuraly equal but their `id_svalue` pointers are different
+       * it may not be safe to use them interchangeably. (TBH I'm not sure if/when
+       * it's safe so I'm erring on the side of caution!) *)
+      phys_equal e1 e2
   | G.NotCst, G.NotCst -> true
   | G.Lit _, _
   | G.Cst _, _
@@ -120,8 +125,11 @@ let union_ctype t1 t2 = if eq_ctype t1 t2 then t1 else G.Cany
 (* aka merge *)
 let union c1 c2 =
   match (c1, c2) with
-  (* For now we only propagate symbolic expressions through linear sequences
-   * of statements. Note that merging symbolic expressions can be tricky.
+  | _any, G.NotCst
+  | G.NotCst, _any ->
+      G.NotCst
+  | c1, c2 when eq c1 c2 -> c1
+  (* Note that merging symbolic expressions can be tricky.
    *
    * Example: Both branches assign `y = x.foo`, but `x` may have a different
    * value in each branch. When merging symbolic expressions like `x.foo` we
@@ -142,11 +150,8 @@ let union c1 c2 =
    *     }
    *)
   | _any, G.Sym _
-  | G.Sym _, _any
-  | _any, G.NotCst
-  | G.NotCst, _any ->
+  | G.Sym _, _any ->
       G.NotCst
-  | c1, c2 when eq c1 c2 -> c1
   | G.Lit l1, G.Lit l2 ->
       let t1 = ctype_of_literal l1 and t2 = ctype_of_literal l2 in
       G.Cst (union_ctype t1 t2)
