+++ tests/rules/sym_prop_merge1.java
@@ -10,10 +10,7 @@ public class LoginServlet extends HttpServlet {
             //ruleid: test
             baz(userName);
         }
-        // TODO: Right now we kill all sym-prop values at JOIN points (i.e. after if-
-        //       or loop statements). In principle we would not have to do this when
-        //       all the branches arriving at the JOIN carry the same sym-prop value.
-        //todoruleid: test
+        //ruleid: test
         baz(userName);
     }
 }
