+++ tests/rules/sym_prop_merge1.yaml
