+++ tests/rules/sym_prop_merge2.java
@@ -1,0 +1,14 @@
+package com.example.log4shell;
+
+public class LoginServlet extends HttpServlet {
+    @Override
+    protected void doPost(HttpServletRequest req) {
+
+        String userName = req.xyz;
+
+        if(foo && x = 42) {
+            //ruleid: test
+            baz(userName);
+        }
+    }
+}
