+++ tests/rules/sym_prop_merge2.yaml
